refactor(client): :sparkles: separate ReviewerDashboard into a new Page component for improved structure
